+++ History.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Add `setHeaders` option
   * Include HTML link in redirect response
   * deps: send@0.5.0
     - Accept string for `maxAge` (converted by `ms`)